add styling to the page

+++ public/app.js
@@ -2,14 +2,16 @@
 
 const ADD_FORM_IDENTIFIER = '#addForm';
 const UPDATE_FORM_IDENTIFIER = '#updateForm';
-const DELETE_CONFIRMATION_IDENTIFIER = '.deleteConfirmation';
 const RESULTS_FROM_REQUEST_IDENTIFIER = '.resultsFromRequest';
 const ADD_BUTTON_IDENTIFIER = '.addButton';
-const LIST_RECORDS_BUTTON_IDENTIFIER = '.listRecordsButton';
+const NEXT_BUTTON_IDENTIFIER = '.nextButton';
+const PREV_BUTTON_IDENTIFIER = '.prevButton';
+const SEARCH_BUTTON_IDENTIFIER = '.searchButton';
 const SHOW_FILTERS_BUTTON_IDENTIFIER = '.showFiltersButton';
 const FILTER_FORM_IDENTIFIER = '#filterForm';
 const MAINTENANCE_LOGGER_URL = '/records';
 const RECORDS_TABLE_IDENTIFIER = '#listTable';
+const RECORDS_PER_PAGE_IDENTIFIER = '#recordsPerPage';
 
 function getMaintenanceRecords(query, callbackFn) {
     $.ajax({
@@ -63,20 +65,16 @@ function getSingleMaintenanceRecords(id, callbackFn) {
 }
 
 function displayMaintenanceRecords(data) {
-    resetScreens();
 
+    $('#listTable tbody').empty();
     $('#listTable tbody').append(
         data.map((record) => { return(`
-            <tr>
+            <tr data-id="${record.id}" >
             <td> ${record.part} </td>
             <td> ${record.status} </td>
             <td> ${record.needsRepair} </td>
             <td> ${new Date(record.lastMaintenance)} </td>
             <td> ${record.frequency} </td>
-            <td> 
-                <button type="button" onclick=viewUpdateRecord("${record.id}")>Update</button> 
-                <button type="button" onclick=viewDeleteRecord("${record.id}")>Delete</button> 
-            </td>
             </tr>
         `)})
     );
@@ -86,79 +84,59 @@ Date.prototype.toDateInputValue = (function() {
     var local = new Date(this);
     local.setMinutes(this.getMinutes() - this.getTimezoneOffset());
     return local.toJSON().slice(0,10);
-});
+}); 
 
-function displayAddRecord(data) {
-    resetScreens();
-    $(ADD_BUTTON_IDENTIFIER).prop("disabled", true);
-    $(ADD_FORM_IDENTIFIER).append(`
-        Part Name: <br> 
-        <input type="text" name="part" placeholder="part name" required> <br>
-        Status: <br> 
-        <input type="text" name="status" placeholder="current status" required> <br>
-        Needs repair?: <br> 
-        <input type="radio" name="needsRepair" value=true> true <br>
-        <input type="radio" name="needsRepair" value=false checked> false <br>
-        Last Maintenance: <br> 
-        <input type="date" name="lastMaintenance" value required> <br>
-        Freq (days): <br> 
-        <input type="number" name="frequency" value="0" required> <br> 
-        <br><br>
-        <input type="submit" value="Add">
-    `);
-    $('input[name="lastMaintenance"]').val(new Date().toDateInputValue());
+function displayFilter(show, instant = false ) {
+    if(show) {
+        $(FILTER_FORM_IDENTIFIER).slideDown( instant ? 0 : 'fast');
+        console.log($(SHOW_FILTERS_BUTTON_IDENTIFIER).attr('background'));
+        $(SHOW_FILTERS_BUTTON_IDENTIFIER).css('background-image', 'url(/filter-up.png)');
+    }
+    else {
+        $(FILTER_FORM_IDENTIFIER).slideUp(instant ? 0 : 'fast');
+        $(SHOW_FILTERS_BUTTON_IDENTIFIER).css('background-image', 'url(/filter-down.png)');
+    }
+}
+
+function displayAddRecord(data) { 
+    $('#addOverlay').css("display","flex");
+    $(ADD_FORM_IDENTIFIER + ' input[name="lastMaintenance"]').val(new Date().toDateInputValue());
+
+    $(ADD_FORM_IDENTIFIER + ' input[name="part"]').focus();
 }
 
 function displayUpdateRecord(data) {
-    resetScreens();
-    $(UPDATE_FORM_IDENTIFIER).append(`
-        Part Name: <br> 
-        <input type="text" name="part" value="${data.part}" required> <br>
-        Status: <br> 
-        <input type="text" name="status" value="${data.status}" required> <br>
-        Repair: <br> 
-        <input type="radio" name="needsRepair" value=true> true <br>
-        <input type="radio" name="needsRepair" value=false> false <br>
-        Last Maintenance: <br> 
-        <input type="date" name="lastMaintenance" value=${data.lastMaintenance.substring(0,10)} required> <br>
-        Frequency: <br> 
-        <input type="number" name="frequency" value=${data.frequency}  required> <br> 
-        id: <br> 
-        <input type="text" name="id" value="${data.id}" readonly> <br> 
-        <br><br>
-        <input type="submit" value="Update">
-    `);
+    $('#updateOverlay').css("display","flex");
+
+    $(UPDATE_FORM_IDENTIFIER + ' input[name="part"]').val(data.part);
+    $(UPDATE_FORM_IDENTIFIER + ' input[name="status"]').val(data.status);
     $(UPDATE_FORM_IDENTIFIER + ' input[name="needsRepair"][value=true]').prop("checked", data.needsRepair);
     $(UPDATE_FORM_IDENTIFIER + ' input[name="needsRepair"][value=false]').prop("checked", !data.needsRepair);
-}
+    $(UPDATE_FORM_IDENTIFIER + ' input[name="lastMaintenance"]').val(data.lastMaintenance.substring(0,10));
+    $(UPDATE_FORM_IDENTIFIER + ' input[name="frequency"]').val(data.frequency);
+    $(UPDATE_FORM_IDENTIFIER + ' input[name="id"]').val(data.id);
+    $('.js-delete-record-btn').attr("id",data.id);
 
-function displayDeleteRecord(data) {
-    resetScreens();
-    $(DELETE_CONFIRMATION_IDENTIFIER).append(`
-        <p>Part Name:  ${data.part} </p>
-        <p>Status: ${data.status} </p>
-        <p>Repair: ${data.needsRepair} </p>
-        <p>Last Maintenance: ${data.lastMaintenance} </p>
-        <p>Freq: ${data.frequency} </p>
-        <br><br><input class="js-delete-record-btn" id= ${data.id}  type="button" value="Delete">
-    `);
+    $(UPDATE_FORM_IDENTIFIER + ' input[name="part"]').focus();
 }
 
 function createFiltersForm(data) {
-    $(FILTER_FORM_IDENTIFIER).hide();
     $(FILTER_FORM_IDENTIFIER).append(`
         <ul>
-        <li><p>Part name: </p><input type="text" name="part" placeholder="all parts"></li>
-        <li><p>Status: </p><input type="text" name="status" placeholder="any status"></li>
-        <li><p>Needs repair: </p><input type="radio" name="needsRepair" value=true>true <input type="radio" name="needsRepair" value=false>false<input type="radio" name="needsRepair" value="any" checked="checked">any</li>
-        <li><p>Begin date: </p><input type="date" name="beginDate" placeholder="begin date"></li>
-        <li><p>End date: </p><input type="date" name="endDate" placeholder="end date"></li>
+        <h2>Filter Search</h2>
+        <li><label>Part name: </label><input type="text" name="part" placeholder="all parts"></li>
+        <li><label>Status: </label><input type="text" name="status" placeholder="any status"></li>
+        <li><label>Needs repair: </label><input type="radio" name="needsRepair" value=true>true <input type="radio" name="needsRepair" value=false>false<input type="radio" name="needsRepair" value="any" checked="checked">any</li>
+        <li><label>Begin date: </label><input type="date" name="beginDate" placeholder="begin date"></li>
+        <li><label>End date: </label><input type="date" name="endDate" placeholder="end date"></li>
+        <button type="button" class="searchButton">Search</button>
         </ul>
     `);
+    displayFilter(false, true);
 }
 
 function displayAddResults(data) {
-    resetScreens();
+    $('#addOverlay').css("display","none");
     $(RESULTS_FROM_REQUEST_IDENTIFIER).empty().append(
         '<p>record added </p>'
     );
@@ -166,7 +144,7 @@ function displayAddResults(data) {
 }
 
 function displayDeleteResults(data) {
-    resetScreens();
+    $('#updateOverlay').css("display","none");
     $(RESULTS_FROM_REQUEST_IDENTIFIER).empty().append(
         '<p>Deleted record </p>'
     );
@@ -174,7 +152,7 @@ function displayDeleteResults(data) {
 }
 
 function displayUpdateResults(data) {
-    resetScreens();
+    $('#updateOverlay').css("display","none");
     $(RESULTS_FROM_REQUEST_IDENTIFIER).empty().append(
         '<p>Updated record</p>'
     );
@@ -187,10 +165,6 @@ function viewAddRecord() {
 
 function viewUpdateRecord(id) {
     getSingleMaintenanceRecords(id, displayUpdateRecord);
-}
-
-function viewDeleteRecord(id) {
-    getSingleMaintenanceRecords(id, displayDeleteRecord);
 }
 
 function getAndDisplayMaintenanceRecords() {
@@ -212,18 +186,14 @@ function getAndDisplayMaintenanceRecords() {
     }
 
     let location = {
-        currentPage: 0,
-        pageQuantity: 20
-    };
-
-    let sort = {
-        part: 1
+        currentPage: $("input[name=currentPage]").val(),
+        pageQuantity: $(RECORDS_PER_PAGE_IDENTIFIER).find(":selected").attr("value")
     };
 
     let query = {
         filter: filter, 
         location: location,
-        sort: sort
+        sort: sortState
     }
 
 	getMaintenanceRecords(query, displayMaintenanceRecords);
@@ -239,15 +209,6 @@ function deleteRecord(id) {
 
 function updateRecord(record) {
     UpdateMaintenanceRecord(record, displayUpdateResults);
-}
-
-function resetScreens() {
-    $(ADD_BUTTON_IDENTIFIER).prop('disabled', false);
-//    $(LIST_RECORDS_BUTTON_IDENTIFIER).prop("disabled", false);
-    $('#listTable tbody').empty();
-    $(UPDATE_FORM_IDENTIFIER).empty();
-    $(ADD_FORM_IDENTIFIER).empty();
-    $(DELETE_CONFIRMATION_IDENTIFIER).empty();
 }
 
 function watchAddBtn() {
@@ -269,21 +230,51 @@ function watchUpdateBtn() {
 }
 
 function watchDeleteBtn() {
-  $(DELETE_CONFIRMATION_IDENTIFIER).on('click', '.js-delete-record-btn', function(event) {
+  $(UPDATE_FORM_IDENTIFIER).on('click', '.js-delete-record-btn', function(event) {
     event.preventDefault();
-    deleteRecord(event.currentTarget.id);
+    if(confirm('Are you sure you want to delete this record?')) deleteRecord(event.currentTarget.id);
   });
 }
 
+function watchCancelBtn() {
+  $(UPDATE_FORM_IDENTIFIER).on('click', '.js-cancel-btn', function(event) {
+    event.preventDefault();
+    $('#updateOverlay').css("display","none");
+  });
+
+  $(ADD_FORM_IDENTIFIER).on('click', '.js-cancel-btn', function(event) {
+    event.preventDefault();
+    $('#addOverlay').css("display","none");
+  });
+}
 
 function watchAddMaintenanceBtn() {
   $(ADD_BUTTON_IDENTIFIER).click((event) => {
+    event.preventDefault();
+    displayFilter(false, true);
     displayAddRecord();
   });
 }
+function watchNextPageBtn() {
+  $(NEXT_BUTTON_IDENTIFIER).click((event) => {
+    event.preventDefault();
+     $(PREV_BUTTON_IDENTIFIER).attr("disabled", false);
+    $("input[name=currentPage]").val(+$("input[name=currentPage]").val()+1);
+    getAndDisplayMaintenanceRecords();
+  });
+}
+function watchPrevPageBtn() {
+  $(PREV_BUTTON_IDENTIFIER).click((event) => {
+    event.preventDefault();
+    let newCurrentPage = +$("input[name=currentPage]").val()-1;
+    $("input[name=currentPage]").val(newCurrentPage);
+    if(newCurrentPage <= 0)  $(PREV_BUTTON_IDENTIFIER).attr("disabled", true);
+    getAndDisplayMaintenanceRecords();
+  });
+}
 
-function watchListRecordsBtn() {
-  $(LIST_RECORDS_BUTTON_IDENTIFIER).click((event) => {
+function watchSearchBtn() {
+  $(FILTER_FORM_IDENTIFIER).on('click', SEARCH_BUTTON_IDENTIFIER, function(event) {
     event.preventDefault();
     getAndDisplayMaintenanceRecords();
   });
@@ -292,24 +283,65 @@ function watchListRecordsBtn() {
 function watchShowFiltersBtn() {
   $(SHOW_FILTERS_BUTTON_IDENTIFIER).click((event) => {
     event.preventDefault();
-    $(FILTER_FORM_IDENTIFIER).toggle('fast');
-    if($(FILTER_FORM_IDENTIFIER).css('display') == 'none') {
-        $(SHOW_FILTERS_BUTTON_IDENTIFIER).text("Show filters");
+    displayFilter($(FILTER_FORM_IDENTIFIER).css('display') == 'none');
+  });
+}
+
+function watchTableRow() {
+  $(RECORDS_TABLE_IDENTIFIER).on('click', "tbody tr", function(event) {
+    viewUpdateRecord($(event.currentTarget).data('id'));
+    event.preventDefault();
+    });
+}
+
+let sortState = {part: 1};
+function watchHeaders() {
+  $('th').click((event) => {
+    event.preventDefault();
+    let sortByField = $(event.currentTarget).data('sort');
+    
+    let sortDirection;
+    if(sortByField in sortState) {
+        sortDirection = sortState[sortByField] * -1; // toggle between 1 & -1 for sort order
     }
     else {
-        $(SHOW_FILTERS_BUTTON_IDENTIFIER).text("Hide filters");
+        sortDirection = 1;
     }
+
+    //clear old sort header's image, then update new sort header's image
+    $(`th[data-sort=${Object.keys(sortState)[0]}] img`).attr("src",""); 
+    $(`th[data-sort=${sortByField}] img`).attr("src", (sortDirection == 1) ? "sort-down.png" : "sort-up.png"); 
+
+    sortState = {};
+    sortState[sortByField] = sortDirection;
+    getAndDisplayMaintenanceRecords();
   });
-};
+}
+
+function watchRecordsPerPage() {
+  $(RECORDS_PER_PAGE_IDENTIFIER).change((event) => {
+    event.preventDefault();
+    $("input[name=currentPage]").val(0),
+    getAndDisplayMaintenanceRecords();
+  });
+}
 
 //  on page load do this
 $(function() {
-    createFiltersForm();
     watchUpdateBtn();
     watchDeleteBtn();
+    watchCancelBtn();
     watchAddBtn();
     watchAddMaintenanceBtn();
-    watchListRecordsBtn();
+    watchSearchBtn();
     watchShowFiltersBtn();
+    watchHeaders();
+    createFiltersForm();
+    watchTableRow();
+    watchRecordsPerPage();
+    watchRecordsPerPage();
+    watchPrevPageBtn();
+    watchNextPageBtn();
+
     getAndDisplayMaintenanceRecords();
 });
